requirements.txt
# ============================================
# DJANGO CORE
# ============================================
Django==5.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# ============================================
# API REST (para integración PHP)
# ============================================
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# ============================================
# FORMULARIOS Y UI
# ============================================
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0

# ============================================
# GENERACIÓN DE PDF
# ============================================
reportlab==4.0.7
pypdf==3.17.4
weasyprint==60.1  # Alternativa a ReportLab
pillow==10.1.0  # Para manejo de imágenes

# ============================================
# PAGOS Y PROCESAMIENTO
# ============================================
stripe==7.8.0  # Para pagos con tarjeta
transbank-sdk==4.0.1  # Para Webpay/OneClick Chile
requests==2.31.0
pycryptodome==3.19.0  # Encriptación

# ============================================
# EXCEL Y REPORTES
# ============================================
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.4  # Para análisis de datos

# ============================================
# TAREAS ASÍNCRONAS
# ============================================
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ============================================
# SEGURIDAD Y AUTENTICACIÓN
# ============================================
django-allauth==0.57.0
django-two-factor-auth==1.15.5
qrcode==7.4.2  # Para 2FA QR codes
pyotp==2.9.0  # Para OTP

# ============================================
# DESARROLLO Y TESTING
# ============================================
django-debug-toolbar==4.2.0
django-extensions==3.2.3
factory-boy==3.3.0
faker==20.1.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# ============================================
# PRODUCCIÓN
# ============================================
gunicorn==21.2.0
whitenoise==6.6.0  # Para servir archivos estáticos
sentry-sdk==1.38.0  # Monitoreo de errores
django-storages==1.14.2  # S3/Cloud storage
boto3==1.34.13  # AWS SDK

# ============================================
# UTILIDADES
# ============================================
python-dateutil==2.8.2
pytz==2023.3
python-slugify==8.0.1
django-environ==0.11.2